requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.13.0=py37h06a4308_0
aiohttp=3.7.4=py37h27cfd23_1
argon2-cffi=20.1.0=py37h27cfd23_1
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.1.0=py_0
blas=2.10=openblas
bleach=3.1.0=py37_0
blinker=1.4=py37h06a4308_0
boost=1.67.0=py37_4
boost-cpp=1.67.0=h14c3975_4
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.6=h14c3975_1002
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
caffe=1.0=py37h2a89de5_5
cairo=1.14.12=h8948797_3
certifi=2022.6.15=py37h06a4308_0
cffi=1.12.3=py37h2e261b9_0
chardet=3.0.4=py37h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
choix=0.3.4=pyhd8ed1ab_1
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.2.1=py_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py37_0
cython=0.29.24=py37h295c915_0
cytoolz=0.9.0.1=py37h14c3975_1001
dask-core=2.0.0=py_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py_0
defusedxml=0.7.1=pyhd3eb1b0_0
dominate=2.6.0=pypi_0
einops=0.3.0=pypi_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=he4413a7_1000
freeglut=3.0.0=hf484d3e_1005
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
geos=3.8.0=he6710b0_0
gettext=0.19.8.1=hc5be6a0_1002
gflags=2.2.2=he1b5a44_1001
giflib=5.1.9=h516909a_0
glib=2.56.2=had28632_1001
glog=0.3.5=hf484d3e_1001
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
graphite2=1.3.13=hf484d3e_1000
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h3010b51_1003
harfbuzz=1.9.0=he243708_1001
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
idna=2.8=pypi_0
imageio=2.8.0=py_0
imgaug=0.4.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2019.4=243
interval=1.0.0=pypi_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.7.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_0
jedi=0.15.1=py37_0
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jovian=0.2.41=pypi_0
jpeg=9c=h14c3975_1001
jsonpatch=1.28=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.0=py_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.6.1=py37_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.1.0=py37he6710b0_0
lame=3.100=h14c3975_1001
leveldb=1.20=h6416369_1001
libblas=3.8.0=10_openblas
libboost=1.67.0=h46d08c1_4
libcblas=3.8.0=10_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1000
libiconv=1.15=h516909a_1005
liblapack=3.8.0=10_openblas
liblapacke=3.8.0=10_openblas
libopenblas=0.3.6=h6e990d7_4
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.9.1=h8b12597_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=h13577e0_0
lmdb=0.9.23=he6710b0_0
lpips=0.1.3=pyhd8ed1ab_0
markdown=3.3.4=py37h06a4308_0
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.1.0=py37h5429711_0
matplotlib-base=3.1.1=py37hfd891ef_0
mistune=0.8.4=py37h14c3975_1001
mkl=2019.4=243
mkl_fft=1.0.13=py37h516909a_1
mkl_random=1.0.4=py37hf2d7682_0
mmdnn=0.2.5=pypi_0
multidict=5.1.0=py37h27cfd23_2
nbconvert=5.5.0=py_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.3=py_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.3.0=py37h06a4308_0
numpy=1.16.4=py37h99e49ec_0
numpy-base=1.16.4=py37h2f8d375_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py37_0
openblas=0.3.6=h6e990d7_4
opencv=3.4.2=py37h6fd60c2_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1q=h7f8727e_0
pandas=0.25.1=py37hb3f55d8_0
pandoc=1.15.0.6=0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.5.1=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_1000
pillow=6.0.0=py37h34e0f95_0
pip=19.1.1=py37_0
pixman=0.38.0=h516909a_1003
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt_toolkit=2.0.9=py_0
protobuf=3.9.1=py37he1b5a44_0
psutil=5.8.0=pypi_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
py-boost=1.67.0=py37h04863e7_4
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.19=py37_0
pygments=2.4.2=py_0
pyjwt=2.1.0=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.0=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
python=3.7.3=h0371630_0
python-dateutil=2.8.0=py37_0
python-gflags=3.1.2=py_0
python-leveldb=0.20=py37hf484d3e_1
python-lmdb=0.96=py37he1b5a44_0
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytube3=9.6.4=pypi_0
pytz=2019.1=py_0
pywavelets=1.0.3=py37hd352d35_1
pyyaml=5.1.1=py37h7b6447c_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.4.3=py37_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.15.0=py37hb3f55d8_2
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.1.0=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=41.0.1=py37_0
shapely=1.7.1=py37h98ec03d_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
snappy=1.1.7=he1b5a44_1002
sqlite=3.28.0=h7b6447c_0
tensorboard=2.1.0=py3_0
tensorboardx=1.8=pypi_0
terminado=0.9.4=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=2.2.0=pyhbf3da8f_0
tk=8.6.9=hed695b0_1002
toolz=0.9.0=py_1
torchfile=0.1.0=pypi_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37h7b6447c_0
tqdm=4.32.2=py_0
traitlets=4.3.2=py37_1000
typing-extensions=3.7.4.1=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.25.3=pypi_0
uuid=1.30=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.7=py_1
webencodings=0.5.1=py37_1
websocket-client=0.56.0=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.33.4=py37_0
widgetsnbextension=3.5.1=py37_0
x264=1!152.20180806=h14c3975_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.9=h516909a_1004
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.7=h14c3975_1000
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
yarl=1.6.3=py37h27cfd23_0
zeromq=4.3.1=he6710b0_3
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0